+++ hypon-cloud/run.sh
@@ -2,6 +2,7 @@
 
 source scripts/home-assistant.sh
 source scripts/hypon.sh
+source scripts/mqtt.sh
 source scripts/variables.sh
 
 loadSensorData() {
@@ -38,7 +39,6 @@ loadSensorData() {
     else
       bashio::log.error "Data Retrieval Error - updating auth token"
       authToken=$(loginHypon)
-      applyTimeModeConfig "$authToken"
     fi
   	sleep "$(bashio::config 'refresh_time')"
   done
@@ -46,5 +46,5 @@ loadSensorData() {
 
 bashio::log.info "Loading Authentication Token"
 authToken=$(loginHypon)
-applyTimeModeConfig "$authToken"
+startMqttControlLoop &
 loadSensorData "$authToken"